Split code into files
This patch splits current ember_csi.py file into smaller files to
make the project cleaner and easier to follow.

This split will also help with the support of multiple CSI spec
versions.
