+++ setup.cfg
@@ -11,7 +11,7 @@ replace = version='{new_version}'
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
 
-[bumpversion:file:ember_csi/ember_csi.py]
+[bumpversion:file:ember_csi/constants.py]
 search = VENDOR_VERSION = '{current_version}'
 replace = VENDOR_VERSION = '{new_version}'
